Put them into their phase file
